chore(app page): move header elements to the Header component chore(auth mixin): move app methods and computed items to the auth mixin
